[KIRA-95] create enums
[KIRA-95] refactor: created enums
