+++ backend/src/api/v1/router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
-from .routes.task_route import router as task_router
+# from .routes.task_route import router as task_router
 
 router = APIRouter(prefix="/kira/app/api/v1")
 
-router.include_router(task_router)
+# router.include_router(task_router)
