+++ backend/src/enums/user_role.py
@@ -1,0 +1,10 @@
+from enum import Enum
+
+class UserRole(str, Enum):
+
+    STAFF = "Staff"
+    MANAGER = "Manager"
+    DIRECTOR = "Director"
+    HR = "HR"
+
+ALLOWED_ROLES = {r.value for r in UserRole}
